requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.0
rouge-score>=0.1.2
accelerate>=0.20.0
opencv-python-headless>=4.0.0